Update whitespace indicators in vim
It looks now like I am showing trailing whitespace ONLY for lines that
I am not editing, which is what I want. Whether or not this is actually
going to work like I want, I'm not sure. But it looks right for now.
Also adds a function to delete trailing whitespace with <leader><space>.
Going to see how I like it.